Handle boundary conditions in separate file
