+++ src/Ising/accessors.cpp
@@ -36,39 +36,6 @@ int Ising::getSize()      {  return this->N;   }
 double Ising::getTemp()   {  return this->T;   }
 
 /**
- * @brief Takes care of the indexing at the boundary of the lattice.
- * 
- * @param ui The row index we "want" to see.
- * @param uj The column index we "want" to see.
- * @param ai The row index the lattice site sees.
- * @param aj The column index the lattice site sees.
- */
-void Ising::BC(int ui, int uj, int* ai, int* aj) {
-	switch (this->boundary) {
-		case BoundaryCondition::PERIODIC:
-			*ai = (ui + this->Ly) % this->Ly; // For periodic boundary condition
-			*aj = (uj + this->Lx) % this->Lx;  //  the lattice is lives on a torus
-			break;
-
-		case BoundaryCondition::SCREW:
-			*ai = (ui + this->Ly) % this->Ly;    // The lattices sites are on a
-			if (uj >= this->Ly || uj < 0) {      // single string. They wrap from
-				*aj = (uj + this->Lx) % this->Lx;  // end of a row to the start of
-				*ai += uj >= this->Lx ? 1 : -1;    // next row
-			}
-			break;
-
-		case BoundaryCondition::FREE:
-			*ai = (ui >= this->Ly || ui < 0) ? -1 : ui; // For free edge boundary
-			*aj = (uj >= this->Lx  || uj < 0) ? -1 : uj; // the site sees nothing
-			break;
-
-		default:
-			std::cout << "Unknown boundary condition" << std::endl;
-	}
-}
-
-/**
  * @brief Lattice point accessor. The index is of the site we "want" to look at.
  * The function takes care of the appropriate boundary conditions and returns
  * the spin value at the actual index in the grid. The function can return `NULL`
@@ -81,7 +48,7 @@ void Ising::BC(int ui, int uj, int* ai, int* aj) {
  */
 bool Ising::operator() (int i, int j) {
 	int ii, jj;
-	this->BC(i, j, &ii, &jj);
+	imposeBC(this->Lx, this-> Ly, i, j, &ii, &jj, this->boundary);
 
 	if (ii == -1 || jj == -1)
 		return NULL;
