+++ src/Ising/dynamics.cpp
@@ -42,8 +42,8 @@ void Ising::flip(int i, int j) {
 void Ising::exchange(int i1, int j1, int i2, int j2) {
 	int i1a, j1a;
 	int i2a, j2a;
-	this->BC(i1, j1, &i1a, &j1a);
-	this->BC(i2, j2, &i2a, &j2a);
+	imposeBC(this->Lx, this->Ly, i1, j1, &i1a, &j1a, this->boundary);
+	imposeBC(this->Lx, this->Ly, i2, j2, &i2a, &j2a, this->boundary);
 
 	bool temp = this->lattice[i1a][j1a];
 	this->lattice[i1a][j1a] = this->lattice[i2a][j2a];
